pyproject.toml
[project]
name = "ems-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "argparse>=1.4.0",
    "dotenv>=0.9.9",
    "openai>=2.21.0",
    "pycryptodome>=3.23.0",
    "pypdf>=6.7.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]
